+++ modules/Student_Billing/MassAssignPayments.php
@@ -17,11 +17,17 @@ if ( $_REQUEST['modfunc'] === 'save' )
 
 				foreach ( (array) $_REQUEST['student'] as $student_id )
 				{
-					$sql .= "INSERT INTO billing_payments (SYEAR,SCHOOL_ID,STUDENT_ID,PAYMENT_DATE,AMOUNT,COMMENTS)
-						VALUES('" . UserSyear() . "',
-						'" . UserSchool() . "','" . $student_id . "','" . $date . "',
-						'" . preg_replace( '/[^0-9.-]/', '', $_REQUEST['amount'] ) . "',
-						'" . $_REQUEST['comments'] . "');";
+					$sql .= DBInsertSQL(
+						'billing_payments',
+						[
+							'SYEAR' => UserSyear(),
+							'SCHOOL_ID' => UserSchool(),
+							'STUDENT_ID' => (int) $student_id,
+							'PAYMENT_DATE' => $date,
+							'AMOUNT' => preg_replace( '/[^0-9.-]/', '', $_REQUEST['amount'] ),
+							'COMMENTS' => $_REQUEST['comments'],
+						]
+					);
 				}
 
 				if ( $sql )